Tried doing some documentation with swagger
