+++ app/src/main/java/genum/app/GenumApplication.java
@@ -1,8 +1,13 @@
 package genum.app;
 
 import genum.product.model.Course;
+import io.swagger.v3.oas.models.ExternalDocumentation;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.info.License;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Import;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
@@ -29,4 +34,19 @@ public class GenumApplication
 
         SpringApplication.run(GenumApplication.class, args);
     }
+
+    @Bean
+    public OpenAPI recommendicOpenApi() {
+        return new OpenAPI()
+                .info(new Info()
+                        .title("genum api")
+                        .description("Backend API for genum")
+                        .version("v0.0.1")
+                        .termsOfService("https://swagger.io/terms/")
+                        .license(new License().name("Apache 2.0").url("https://springdoc.org")))
+                .externalDocs(new ExternalDocumentation()
+                        .description("SpringShop Wiki Documentation")
+                        .url("https://springshop.wiki.github.org/docs"));
+    }
+
 }
