+++ app/src/main/resources/application.properties
@@ -25,6 +25,6 @@ google.credentials=${GOOGLE_CREDENTIALS:""}
 
 cors.frontend_domain=${FRONTEND_DOMAIN:""}
 
-logging.level.org.springframework.security=trace
+logging.level.org.springframework.security=debug
 logging.level.org.springframework.security.config.annotation.authentication.configuration.InitializeUserDetailsBeanManagerConfigurer=error
