+++ email/pom.xml
@@ -51,6 +51,7 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
             </plugin>
+
         </plugins>
     </build>
