+++ shared/src/main/java/genum/shared/config/CommonConfig.java
@@ -1,9 +1,5 @@
 package genum.shared.config;
 
-import io.swagger.v3.oas.models.ExternalDocumentation;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.info.License;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.MongoDatabaseFactory;
@@ -29,19 +25,5 @@ public class CommonConfig {
     @Bean
     public RestTemplate restTemplate(){
         return new RestTemplate();
-    }
-
-    @Bean
-    public OpenAPI recommendicOpenApi() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("genum api")
-                        .description("Backend API for genum")
-                        .version("v0.0.1")
-                        .termsOfService("https://swagger.io/terms/")
-                        .license(new License().name("Apache 2.0").url("https://springdoc.org")))
-                .externalDocs(new ExternalDocumentation()
-                        .description("SpringShop Wiki Documentation")
-                        .url("https://springshop.wiki.github.org/docs"));
     }
 }
